+++ Tests/KeyValueCRDTTests/UIKeyValueDocumentTests.swift
@@ -17,7 +17,7 @@ final class UIKeyValueDocumentTests: XCTestCase {
     }
     waitForExpectations(timeout: 3)
     XCTAssertFalse(document.hasUnsavedChanges)
-    try document.keyValueCRDT.writeText("Hello, world", to: "test")
+    try document.keyValueCRDT?.writeText("Hello, world", to: "test")
     XCTAssertTrue(document.hasUnsavedChanges)
     let closedExpectation = expectation(description: "closed")
     document.close { success in
@@ -33,7 +33,7 @@ final class UIKeyValueDocumentTests: XCTestCase {
       rtOpen.fulfill()
     }
     waitForExpectations(timeout: 3)
-    let result = try roundTrip.keyValueCRDT.read(key: "test")
+    let result = try roundTrip.keyValueCRDT!.read(key: "test")
     XCTAssertEqual(try result.text, "Hello, world")
     XCTAssertFalse(roundTrip.hasUnsavedChanges)
     let rtClose = expectation(description: "round trip close")